# Optimise use of sub-factories in tests