--- a/datahub/investment/project/report/test/factories.py
+++ b/datahub/investment/project/report/test/factories.py
@@ -16,7 +16,7 @@ class SPIReportFactory(factory.django.DjangoModelFactory):
 
     created_on = now()
     created_by = factory.SubFactory(AdviserFactory)
-    modified_by = factory.SubFactory(AdviserFactory)
+    modified_by = factory.SelfAttribute('created_by')
 
     class Meta:
         model = SPIReport