--- a/datahub/omis/quote/test/factories.py
+++ b/datahub/omis/quote/test/factories.py
@@ -12,7 +12,7 @@ class QuoteFactory(factory.django.DjangoModelFactory):
 
     id = factory.LazyFunction(uuid.uuid4)
     created_by = factory.SubFactory(AdviserFactory)
-    modified_by = factory.SubFactory(AdviserFactory)
+    modified_by = factory.SelfAttribute('created_by')
     reference = factory.Faker('text', max_nb_chars=10)
     content = factory.Faker('text')
     expires_on = factory.Faker('future_date')
@@ -26,7 +26,7 @@ class CancelledQuoteFactory(QuoteFactory):
     """Cancelled Order factory."""
 
     cancelled_on = now()
-    cancelled_by = factory.SubFactory(AdviserFactory)
+    cancelled_by = factory.SelfAttribute('created_by')
 
 
 class AcceptedQuoteFactory(QuoteFactory):